refactor: refactoring

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -24,7 +24,7 @@ public class BasicBinaryContentService implements BinaryContentService {
         BinaryContent binaryContent =
             new BinaryContent(
                 fileName,
-                (long) bytes.length,
+                bytes.length,
                 contentType,
                 bytes
             );
