+++ src/main/java/com/sprint/mission/discodeit/service/file/FileBinaryContentService.java
@@ -24,7 +24,7 @@ public class FileBinaryContentService implements BinaryContentService {
         BinaryContent binaryContent =
             new BinaryContent(
                 fileName,
-                (long) bytes.length,
+                bytes.length,
                 contentType,
                 bytes
             );
