+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFUserService.java
@@ -53,7 +53,7 @@ public class JCFUserService implements UserService {
         String fileName = binaryContentRequest.getName();
         String contentType = binaryContentRequest.getContentType();
         byte[] bytes = binaryContentRequest.getBytes();
-        long size = bytes.length;
+        int size = bytes.length;
 
         BinaryContent binaryContent = new BinaryContent(fileName, size, contentType, bytes);
         binaryContentRepository.save(binaryContent);
@@ -116,7 +116,7 @@ public class JCFUserService implements UserService {
         String fileName = binaryContentRequest.getName();
         String contentType = binaryContentRequest.getContentType();
         byte[] bytes = binaryContentRequest.getBytes();
-        long length = bytes.length;
+        int length = bytes.length;
 
         BinaryContent binaryContent = new BinaryContent(fileName,
             length,
@@ -124,7 +124,7 @@ public class JCFUserService implements UserService {
             bytes);
 
         user.updateProfileImageId(binaryContent.getId());
-        
+
         return user.getId();
     }
